Simplified self_registration for non neuanmeldungen
fixes #162
