+++ app/views/groups/self_registration/_main_person_simple.html.haml
@@ -1,0 +1,13 @@
+-#  Copyright (c) 2023, Schweizer Alpen-Club. This file is part of
+-#  hitobito_sac_cas and licensed under the Affero General Public License version 3
+-#  or later. See the COPYING file at the top-level directory or at
+-#  https://github.com/hitobito/hitobito_sac_cas.
+
+- if f.object.is_a?(SelfRegistration::MainPerson)
+  = f.labeled(:newsletter, '&nbsp;'.html_safe) do
+    %label.checkbox
+      = f.check_box :newsletter
+      = t('groups.self_registration.summary.newsletter')
+
+  = field_set_tag do
+    = f.labeled_input_field(:promocode)
